+++ temba/utils/analytics/base.py
@@ -1,17 +1,15 @@
 import abc
 import logging
 
-from django.conf import settings
+from django.template import Context, Engine
 from django.utils.safestring import mark_safe
 
 logger = logging.getLogger(__name__)
 
-registered_backends = []
-
 
 class AnalyticsBackend(metaclass=abc.ABCMeta):
     slug: str = None
-    template_hooks = {}
+    hook_templates = {}
 
     def gauge(self, event: str, value):
         """
@@ -33,38 +31,23 @@ class AnalyticsBackend(metaclass=abc.ABCMeta):
         Notifies of a user's consent status.
         """
 
-    def get_template_html(self, hook: str) -> str:
+    def get_hook_template(self, name: str) -> str:
         """
-        Gets HTML to be inserted at the named template hook
+        Gets template name for named hook
         """
+        return self.hook_templates.get(name)
 
-
-def init():  # pragma: no cover
-    """
-    Initializes our analytics backends based on our settings.
-    TODO this should all be dynamic based on a setting which is a list of backend class names
-    """
-
-    from .crisp import CrispBackend
-    from .librato import LibratoBackend
-
-    registered_backends.clear()
-
-    # configure Librato if configured
-    librato_user = getattr(settings, "LIBRATO_USER", None)
-    librato_token = getattr(settings, "LIBRATO_TOKEN", None)
-    if librato_user and librato_token:
-        registered_backends.append(LibratoBackend(librato_user, librato_token))
-
-    crisp_identifier = getattr(settings, "CRISP_IDENTIFIER", None)
-    crisp_key = getattr(settings, "CRISP_KEY", None)
-    crisp_website_id = getattr(settings, "CRISP_WEBSITE_ID", None)
-    if crisp_identifier and crisp_key and crisp_website_id:
-        registered_backends.append(CrispBackend(crisp_identifier, crisp_key, crisp_website_id))
+    def get_hook_context(self, request) -> dict:
+        """
+        Gets context to be included in hook templates
+        """
+        return {}
 
 
 def get_backends() -> list:
-    return registered_backends
+    from . import backends
+
+    return list(backends.values())
 
 
 def gauge(event: str, value):
@@ -115,14 +98,18 @@ def track(user, event: str, properties: dict = None):
             logger.exception(f"error tracking event on {backend.slug}")
 
 
-def get_template_html(hook: str) -> str:
+def get_hook_html(name: str, context) -> str:
     """
     Gets HTML to be inserted at the named template hook
     """
-    content = ""
+    engine = Engine.get_default()
+    html = ""
     for backend in get_backends():
-        html = backend.get_template_html(hook)
-        if html:
-            content += f"<!-- begin hook for {backend.slug} -->\n{html}\n<!-- end hook for {backend.slug} -->\n"
+        template_name = backend.get_hook_template(name)
+        if template_name:
+            with context.update(backend.get_hook_context(context.request)):
+                html += f"<!-- begin hook for {backend.slug} -->\n"
+                html += engine.get_template(template_name).render(Context(context))
+                html += f"<!-- end hook for {backend.slug} -->\n"
 
-    return mark_safe(content)
+    return mark_safe(html)