+++ temba/utils/analytics/crisp/tests.py
@@ -2,11 +2,14 @@ import random
 from unittest import mock
 from unittest.mock import MagicMock
 
+from django.test.utils import override_settings
+
 from temba.tests import TembaTest
 
 from .backend import CrispBackend
 
 
+@override_settings(CRISP_IDENTIFIER="CI123", CRISP_KEY="CK234", CRISP_WEBSITE_ID="CW345")
 class CrispTest(TembaTest):
     def setUp(self):
         super().setUp()
@@ -15,7 +18,7 @@ class CrispTest(TembaTest):
 
         self.mock_website = MagicMock()
 
-        self.backend = CrispBackend("acme", "sesame", "my_site")
+        self.backend = CrispBackend()
         self.backend.client.website = self.mock_website
 
     def test_gauge(self):
@@ -26,21 +29,21 @@ class CrispTest(TembaTest):
         self.backend.track(self.user, "user_signup", {"user_id": "123", "nick_name": None})
 
         self.mock_website.add_people_event.assert_called_with(
-            "my_site", "User@nyaruka.com", {"color": "green", "text": "user_signup", "data": {"user_id": "123"}}
+            "CW345", "User@nyaruka.com", {"color": "green", "text": "user_signup", "data": {"user_id": "123"}}
         )
 
         # created events in blue...
         self.backend.track(self.admin, "foo_created", {"foo_id": "234"})
 
         self.mock_website.add_people_event.assert_called_with(
-            "my_site", "Administrator@nyaruka.com", {"color": "blue", "text": "foo_created", "data": {"foo_id": "234"}}
+            "CW345", "Administrator@nyaruka.com", {"color": "blue", "text": "foo_created", "data": {"foo_id": "234"}}
         )
 
         # export events in purple...
         self.backend.track(self.admin, "foo_export", {"foo_id": "345"})
 
         self.mock_website.add_people_event.assert_called_with(
-            "my_site",
+            "CW345",
             "Administrator@nyaruka.com",
             {"color": "purple", "text": "foo_export", "data": {"foo_id": "345"}},
         )
@@ -53,7 +56,7 @@ class CrispTest(TembaTest):
 
         # did we actually call save?
         self.mock_website.add_new_people_profile.assert_called_with(
-            "my_site",
+            "CW345",
             {
                 "person": {"nickname": " "},
                 "company": {
@@ -73,7 +76,7 @@ class CrispTest(TembaTest):
         self.backend.identify(self.admin, {"slug": "test", "host": "rapidpro.io"}, self.org)
 
         self.mock_website.update_people_profile.assert_called_with(
-            "my_site",
+            "CW345",
             self.admin.email,
             {
                 "person": {"nickname": " "},
@@ -93,7 +96,7 @@ class CrispTest(TembaTest):
         self.backend.change_consent(self.admin, True)
 
         self.mock_website.update_people_profile.assert_called_with(
-            "my_site", "Administrator@nyaruka.com", {"segments": ["consented"]}
+            "CW345", "Administrator@nyaruka.com", {"segments": ["consented"]}
         )
 
         # valid user which did not consent
@@ -103,12 +106,21 @@ class CrispTest(TembaTest):
         self.backend.change_consent(self.admin, False)
 
         self.mock_website.save_people_data.assert_called_with(
-            "my_site", self.admin.email, {"data": {"consent_changed": mock.ANY}}
+            "CW345", self.admin.email, {"data": {"consent_changed": mock.ANY}}
         )
 
-    def test_get_template_html(self):
+    def test_get_hook_template(self):
+        self.assertEqual("utils/analytics/crisp/login.html", self.backend.get_hook_template("login"))
+
+        request = MagicMock()
+        request.user = self.admin
+
+        self.assertEqual({"crisp_website_id": "CW345", "crisp_token_id": None}, self.backend.get_hook_context(request))
+
+        user_settings = self.admin.get_settings()
+        user_settings.external_id = "AD567"
+        user_settings.save()
+
         self.assertEqual(
-            '<script type="text/javascript">$crisp.push(["do", "session:reset"])</script>',
-            self.backend.get_template_html("login"),
+            {"crisp_website_id": "CW345", "crisp_token_id": "AD567"}, self.backend.get_hook_context(request)
         )
-        self.assertEqual("", self.backend.get_template_html("foo"))