+++ temba/utils/analytics/tests.py
@@ -1,6 +1,8 @@
 from unittest.mock import MagicMock, patch
 
 from django.contrib.auth.models import AnonymousUser
+from django.template import Engine, Template
+from django.urls import reverse
 
 from temba.tests import TembaTest
 from temba.utils import analytics
@@ -55,26 +57,40 @@ class AnalyticsTest(TembaTest):
         good.change_consent.assert_called_once_with(self.user, True)
 
     @patch("temba.utils.analytics.base.get_backends")
-    def test_get_template_html(self, mock_get_backends):
+    def test_get_hook_html(self, mock_get_backends):
         good = MagicMock()
         good.slug = "good"
-        good.get_template_html.return_value = '<script>alert("good")</script>'
+        good.get_hook_template.return_value = "good/frame_top.html"
         mock_get_backends.return_value = [BadBackend(), good]
 
-        self.assertEqual(
-            """<!-- begin hook for bad -->
+        real_get_template = Engine.get_default().get_template
+
+        def get_template(name):
+            if name == "good/frame_top.html":
+                return Template('<script>alert("good")</script>\n')
+            elif name == "bad/frame_top.html":
+                return Template('<script>alert("bad")</script>\n')
+            else:
+                return real_get_template(name)
+
+        with patch("django.template.engine.Engine.get_template", wraps=get_template):
+            self.login(self.admin)
+            response = self.client.get(reverse("orgs.org_home"))
+
+            self.assertContains(
+                response,
+                """<!-- begin hook for bad -->
 <script>alert("bad")</script>
 <!-- end hook for bad -->
 <!-- begin hook for good -->
 <script>alert("good")</script>
-<!-- end hook for good -->
-""",
-            analytics.get_template_html("login"),
-        )
+<!-- end hook for good -->""",
+            )
 
 
 class BadBackend(AnalyticsBackend):
     slug = "bad"
+    hook_templates = {"frame-top": "bad/frame_top.html"}
 
     def gauge(self, event: str, value):
         raise ValueError("boom")
@@ -87,6 +103,3 @@ class BadBackend(AnalyticsBackend):
 
     def change_consent(self, user, consent: bool):
         raise ValueError("boom")
-
-    def get_template_html(self, hook) -> str:
-        return '<script>alert("bad")</script>'